requirements.txt
praw~=7.8.1
pandas~=2.2.3
python-dotenv~=1.0.1
pytrends~=4.9.2
matplotlib~=3.10.0
wordcloud~=1.9.4
seaborn~=0.13.2